extract and reuse course admin actions
