+++ resources/js/Pages/Courses/components/CourseItemCard.vue
@@ -1,21 +1,17 @@
 <script setup>
-import { Link, router, usePage } from '@inertiajs/vue3';
+import { Link, usePage } from '@inertiajs/vue3';
 import { computed } from 'vue';
-import { useI18n } from 'vue-i18n';
+import CourseAdminActions from './CourseAdminActions.vue';
 
-const { t } = useI18n();
 const page = usePage();
 const isAdmin = computed(() => page.props.auth.isAdmin);
 
 const { course } = defineProps({
-  course: Object,
+  course: {
+    type: Object,
+    required: true,
+  },
 });
-
-const deleteCourse = (id) => {
-  if (confirm(t('courses.delete_confirm'))) {
-    router.delete(`/courses/${id}`);
-  }
-};
 </script>
 
 <template>
@@ -28,26 +24,12 @@ const deleteCourse = (id) => {
             class="card-title fs-4 stretched-link"
             >{{ course.name }}</Link
           >
-          <p class="card-text text-muted">{{ course.school_year }}</p>
+          <p class="card-text text-muted">
+            {{ course.school_year }}
+            <span v-if="course.description"> | {{ course.description }}</span>
+          </p>
         </div>
-        <div v-if="isAdmin" class="col-auto">
-          <div class="btn-group">
-            <Link
-              class="btn btn-outline-secondary"
-              :title="$t('courses.edit')"
-              :href="`/courses/${course.id}/edit`"
-            >
-              <i class="ri ri-pencil-line"></i>
-            </Link>
-            <button
-              class="btn btn-outline-danger"
-              :title="$t('courses.delete')"
-              @click="deleteCourse(course.id)"
-            >
-              <i class="ri ri-delete-bin-line"></i>
-            </button>
-          </div>
-        </div>
+        <CourseAdminActions v-if="isAdmin" class="col-auto" :course="course" />
       </div>
     </div>
   </div>
